+++ security/samsung/defex_lsm/defex_catch_list_compat.c
@@ -24,7 +24,7 @@ const int defex_nr_syscalls_compat = DEFEX_CATCH_COUNT;
 
 const struct local_syscall_struct *get_local_syscall_compat(int syscall_no)
 {
-	if (syscall_no >= __NR_compat_syscalls)
+	if ((unsigned int)syscall_no >= __NR_compat_syscalls)
 		return NULL;
 
 	if (!syscall_catch_arr[syscall_no].local_syscall && !syscall_catch_arr[syscall_no].err_code && syscall_no) {
